Restructure product
